% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_sequence.R
\name{annot_sequence}
\alias{annot_sequence}
\title{Modify annotated sequence with indistinguishable residue positions.}
\usage{
annot_sequence(seq, residue_posn, pattern)
}
\arguments{
\item{seq}{A character vector representing amino acid sequence.}

\item{residue_posn}{A character vector specifying the positions of residues to annotates.}

\item{pattern}{A character vector specifying the pattern of the residue positions: use "positions not distinguishable" for this function.}
}
\value{
A character vector representing the modified annotated sequence.
}
\description{
This function modifies an annotated sequence when residue positions are not distinguishable.
}
\details{
This function checks if the residue positions in the sequence are not distinguishable.
If the positions are not distinguishable, it modifies the annotated sequence (results from the annotate_positions function) accordingly.
Specifically, it inserts '[+80]' after the first occurrence of 'S', 'T', or 'Y' if it is not already annotated with '[+80]'.
If there are multiple occurrences of 'S', 'T', or 'Y' and the first occurrence coincides with a position already annotated with '[+80]',
it inserts '[+80]' after the next occurrence of 'S', 'T', or 'Y'.
}
\examples{
seq <- "INSAESMELWTSYQK"
residue_posn <- "S; S6"
pattern <- "positions not distinguishable"
annot_sequence(seq, residue_posn, pattern)
# [1] "INS[+80]AES[+80]MELWTSYQK"

seq <- "INSAESMELWTSYQK"
residue_posn <- "S; S3"
pattern <- "positions not distinguishable"
annot_sequence(seq, residue_posn, pattern)
# [1] "INS[+80]AES[+80]MELWTSYQK"


seq <- "INELEQSINDLRAEMGVEGTPPPASK"
residue_posn <- "M15; T/S"
pattern <- "positions not distinguishable"
annot_sequence(seq, residue_posn, pattern)
# [1] "INELEQS[+80]INDLRAEM[+16]GVEGTPPPASK"

seq <- "KTDGSTTPAYAHGQHHSIFSPATGAVSDSSLK"
residue_posn <- "Y/T/S"
pattern <- "positions not distinguishable"
annot_sequence(seq, residue_posn, pattern)
# [1] "KT[+80]DGSTTPAYAHGQHHSIFSPATGAVSDSSLK"


}
\seealso{
\code{\link{annotate_positions}}, \code{\link{regexpr}}, \code{\link{substr}}
}
