% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_positions.R
\name{annotate_positions}
\alias{annotate_positions}
\title{Annotate position in a sequence}
\usage{
annotate_positions(seq, residue_posn)
}
\arguments{
\item{seq}{A character vector representing amino acid sequence.}

\item{residue_posn}{A character vector specifying the positions of residues to annotates.}
}
\value{
A single character vector representing the annotated sequence.
}
\description{
This function annotates residues in a given sequence with additional characters. This code is executed when the residue position is distinguishable.
}
\details{
This function takes a sequence and a vector of residue positions and annotates those positions with specific characters.
The data needs to be cleaned to some extent so that the residues in residues_posn should only be 'M', 'S', 'T', or 'Y'.
Each of these residues will be annotated with specific strings: 'S' with '[+80]', 'M' with '[+16]', 'T' with '[+80]', and 'Y' with '[+80]'.
The addition of  80 dalton mass [+80] to S, Y, and T indicates phosphorylation while addition of [+16] to M indicates oxidation.
The annotated sequence will be returned with the inserted strings at the respective positions.
}
\examples{
seq <- "MSTY"
residue_posn <- "M1; S2; T3; Y4"
annotate_positions(seq, residue_posn)
# [1] "M[+16]S[+80]T[+80]Y[+80]"

}
\seealso{
\code{\link{grepl}}, \code{\link[stringr]{str_match_all}}
}
